move shared types into a dedicated types-folder
