+++ apps/server/src/terrain/types/basetypes.ts
@@ -1,0 +1,4 @@
+export interface PositionData {
+    latitude: number;
+    longitude: number;
+}
