+++ apps/server/src/terrain/types/elevationgrid.ts
