+++ apps/server/src/terrain/types/processing.ts
@@ -1,0 +1,6 @@
+export interface ElevationProfile {
+    pathWidth: number,
+    waypointsLatitudes: number[],
+    waypointsLongitudes: number[],
+    range: number,
+}
